refactor: 상품 목록 화면 코드 가독성 개선
OnScrollListener 파일 분리
더보기 로직 Interface 정의
