+++ app/src/main/java/woowacourse/shopping/view/main/ProductsOnScrollListener.kt
@@ -1,0 +1,29 @@
+package woowacourse.shopping.view.main
+
+import android.view.View
+import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import woowacourse.shopping.databinding.ActivityMainBinding
+
+class ProductsOnScrollListener(
+    private val binding: ActivityMainBinding,
+    private val viewModel: ProductsViewModel,
+) : RecyclerView.OnScrollListener() {
+    override fun onScrolled(
+        recyclerView: RecyclerView,
+        dx: Int,
+        dy: Int,
+    ) {
+        val layoutManager = recyclerView.layoutManager as GridLayoutManager
+        val lastVisibleItemPosition = layoutManager.findLastVisibleItemPosition()
+        binding.btnLoadMoreProducts.visibility =
+            if (
+                lastVisibleItemPosition == layoutManager.itemCount - 1 &&
+                viewModel.totalSize > viewModel.products.size
+            ) {
+                View.VISIBLE
+            } else {
+                View.GONE
+            }
+    }
+}
