+++ app/src/main/res/layout/activity_main.xml
@@ -6,6 +6,9 @@
         <variable
             name="viewModel"
             type="woowacourse.shopping.view.main.ProductsViewModel" />
+        <variable
+            name="handler"
+            type="woowacourse.shopping.view.main.ProductsEventHandler" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
@@ -37,7 +40,7 @@
             android:backgroundTint="@color/pagination_enabled"
             android:insetTop="0dp"
             android:insetBottom="0dp"
-            android:onClick="@{()->viewModel.requestProductsPage(viewModel.pageLiveData + 1)}"
+            android:onClick="@{()->handler.onLoadMoreProducts(viewModel.pageLiveData + 1)}"
             android:text="@string/button_text_load_more"
             android:textSize="20sp"
             android:textStyle="bold"
